+++ apps/astarte_appengine_api/lib/astarte_appengine_api_web/endpoint.ex
@@ -42,7 +42,7 @@ defmodule Astarte.AppEngine.APIWeb.Endpoint do
   end
 
   plug Plug.RequestId
-  plug Plug.Logger
+  plug PlugLoggerWithMeta
 
   plug Plug.Parsers,
     parsers: [:urlencoded, :multipart, :json],